Rewrite initial state annotations generation
Fix useValues types
